Lectures/3EloadasPythonKodok-20251104/requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycryptodome==3.23.0
Werkzeug==3.1.3
